+++ docs/whats_new.rst
@@ -34,6 +34,7 @@ Enhancements
 - Including missed tag in all the models and creating some test for it (:gh:`811` by `Bruno Aristimunha`_ )
 - Added :class:`braindecode.models.LUNA`  (:gh:`802` by `Bruno Aristimunha`_, `Thorir Mar Ingolfsson`_)
 - Modify verbosity of create_window_from_events (:gh:`814` by `Matthew Chen`_)
+- Modify the interface of FBMSNet for consistency with FBCNet + Unit Tests (:gh:`819` by `Matthew Chen`_)
 
 
 API changes